Scorecard/requirements.txt
requests>=2.25.1
pyodbc>=4.0.30
python-dotenv>=0.19.0
pandas>=1.3.0
selenium>=4.0.0
openpyxl>=3.0.9
xlrd>=2.0.1 